rocr: Add AMD ext for configuring a queue's HW context
This adds an AMD extension API for configuring a queue's
hardware context. For AIEs the embedded runtime (ERT)
creates a hardware context for a queue that has various
configurable parameters, notably CUs can be configured.

This API could also be used for future GPU HW context
configuration (e.g., it could be used as a general way
to set the CU masks).

Change-Id: I7d1ce36d7ad6830f1b0c867bb2646e6e09cf4845